pyproject.toml
[project]
name = "google_calendar_mcp"
version = "0.1.0"
description = "Google Calendar MCP server."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "mcp>=1.15.0",
    "python-dotenv>=1.1.1",
    "cryptography>=46.0.2",
    "gunicorn>=23.0.0",
    "redis>=6.4.0",
    "jotsu-mcp>=0.14.0",
    "python-dateutil>=2.9.0.post0",
    "google-auth>=2.41.1",
    "google-api-python-client>=2.184.0",
]

[dependency-groups]
dev = [
    "click>=8.3.0",
    "ruff>=0.13.2",
]
